Reduced size of images displayed in team list
The team list will now display a low resolution version of the team photo. Clicking it will lead you to the high res image in a new tab.
Reduces loading time of the website.
Had to disable the external website icon as it is not compatible to using target with an image.